Extracting time step calculation function from env to the specific module.
